Moved LookupParams to a separate file.
